adding loaderItems to ItemMenu on OldPathNavigator
