+++ static-assets/components/cstudio-contextual-nav/wcm-site-dropdown-mods/wcm-assets-folder.js
@@ -589,6 +589,10 @@ var storage = CStudioAuthoring.Storage;
 
         const path = treeNodeTO.uri;
 
+        const loaderItems = CrafterCMSNext.util.content.getNumOfMenuOptionsForItem(
+          CrafterCMSNext.util.content.parseLegacyItemToDetailedItem(treeNodeTO)
+        );
+
         const openItemMenu = () => {
           CrafterCMSNext.system.store.dispatch({
             type: 'BATCH_ACTIONS',
@@ -609,6 +613,7 @@ var storage = CStudioAuthoring.Storage;
                 type: 'SHOW_ITEM_MENU',
                 payload: {
                   path,
+                  loaderItems,
                   anchorReference: 'anchorPosition',
                   anchorPosition: { top: event.clientY - 10, left: event.clientX - 10 }
                 }
@@ -623,7 +628,6 @@ var storage = CStudioAuthoring.Storage;
         menuIcon.onclick = function(event) {
           event.preventDefault();
           event.stopPropagation();
-          const path = treeNodeTO.uri;
           openItemMenu();
         };
         nodeSpan.appendChild(menuIcon);
