+++ static-assets/components/cstudio-contextual-nav/wcm-site-dropdown-mods/wcm-root-folder.js
@@ -995,6 +995,10 @@
 
           const path = treeNodeTO.uri;
 
+          const loaderItems = CrafterCMSNext.util.content.getNumOfMenuOptionsForItem(
+            CrafterCMSNext.util.content.parseLegacyItemToDetailedItem(treeNodeTO)
+          );
+
           const openItemMenu = () => {
             CrafterCMSNext.system.store.dispatch({
               type: 'BATCH_ACTIONS',
@@ -1015,6 +1019,7 @@
                   type: 'SHOW_ITEM_MENU',
                   payload: {
                     path,
+                    loaderItems,
                     anchorReference: 'anchorPosition',
                     anchorPosition: { top: event.clientY - 10, left: event.clientX - 10 }
                   }
