+++ ui/app/src/components/Navigation/PathNavigator/utils.ts
@@ -15,7 +15,6 @@
  */
 
 import { DetailedItem } from '../../../models/Item';
-import { isRootPath } from '../../../utils/path';
 
 export function isNavigable(item: DetailedItem): boolean {
   return item.systemType === 'page';
@@ -35,15 +34,6 @@ export function isFolder(item: DetailedItem): boolean {
   return item.systemType === 'folder';
 }
 
-export function getNumOfMenuOptionsForItem(item: DetailedItem): number {
-  if (isNavigable(item)) {
-    return isRootPath(item.path) ? 11 : 16;
-  } else if (isFolder(item)) {
-    return isRootPath(item.path) ? 3 : 6;
-  } else if (isPreviewable(item)) {
-    return 10;
-  }
-}
 export function rand(min: number, max: number) {
   return Math.floor(Math.random() * (max - min + 1) + min);
 }
