+++ ui/app/src/components/SystemStatus/GlobalDialogManager.tsx
@@ -405,6 +405,7 @@ function GlobalDialogManager() {
       <ItemMenu
         open={state.itemMenu.open}
         path={state.itemMenu.path}
+        loaderItems={state.itemMenu.loaderItems}
         onClose={createCallback(state.itemMenu.onClose, dispatch)}
         anchorReference={state.itemMenu.anchorReference}
         anchorPosition={state.itemMenu.anchorPosition}
